feat(ItemList): move itemListItem to seperate folder
