+++ src/components/molecules/ItemList/ItemListItem/ItemListItem.scss
@@ -1,0 +1,34 @@
+@use "../../../../assets/styles/mixins";
+
+.itemListItem {
+    display: flex;
+    align-items: center;
+    gap: var(--guit-ref-spacing-xsmall);
+    width: 100%;
+    min-height: 3.2rem;
+    padding-block: var(--guit-ref-spacing-2xsmall);
+    padding-inline: var(--guit-ref-spacing-medium);
+    user-select: none;
+    border-radius: var(--guit-ref-radius-3xsmall);
+    background-color: var(--guit-sem-color-background-neutral-1-nudge);
+    color: var(--guit-sem-color-foreground-neutral-2);
+
+    &_selected {
+        color: var(--guit-sem-color-foreground-selected);
+    }
+
+    &_disabled {
+        pointer-events: none;
+        color: var(--guit-sem-color-foreground-disabled);
+    }
+
+    @include mixins.media-hover {
+        &:hover {
+            background-color: var(--guit-sem-color-background-neutral-1-nudge-hover);
+        }
+    }
+
+    &:active {
+        background-color: var(--guit-sem-color-background-neutral-1-nudge-pressed);
+    }
+}
