+++ src/components/molecules/ItemList/index.tsx
@@ -1,3 +1,3 @@
 export { IItemListProps, default } from "./ItemList";
 export { IItemListFooterProps, default as ItemListFooter } from "./ItemListFooter";
-export { IItemListItemProps, default as ItemListItem } from "./ItemListItem";
+export { IItemListItemProps, default as ItemListItem } from "./ItemListItem/ItemListItem";
